Rename factories/index.js to factories/dashboard.js
